requirements.txt
# requirements.txt - Vortex Board
# Versões fixadas para ambiente de produção

# === DJANGO CORE ===
Django==4.2.8
django-environ==0.11.2

# === DATABASE ===
psycopg2-binary==2.9.9

# === ASYNC & WEBSOCKETS ===
channels[daphne]==4.0.0
channels-redis==4.1.0
uvicorn[standard]==0.24.0

# === FRONTEND INTEGRATION ===
django-htmx==1.17.2
django-tailwind==3.6.0

# === REPORTS & FILES ===
reportlab==4.0.7
xlsxwriter==3.1.9
openpyxl==3.1.2
Pillow==10.1.0

# === DEVELOPMENT TOOLS ===
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# === PRODUCTION SERVER ===
gunicorn==21.2.0
whitenoise==6.6.0

# === TESTING ===
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
factory-boy==3.3.0

# === CODE QUALITY ===
black==23.11.0
flake8==6.1.0
isort==5.12.0

# === MONITORING & LOGS ===
structlog==23.2.0
django-health-check==3.17.0

# === TASK QUEUE (OPCIONAL) ===
celery==5.3.4
redis==5.0.1

# === API & SERIALIZATION ===
djangorestframework==3.14.0
django-cors-headers==4.3.1

# === UTILS ===
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# === SECURITY ===
django-csp==3.7
django-ratelimit==4.1.0

# === DEV DEPENDENCIES (comentar em produção) ===
# django-model2puml==0.2.2
# ipython==8.17.2
# jupyter==1.0.0